refactor: rename types to schema architecture
- Move src/types/ to src/schema/ for better naming consistency
- Move test/types/ to test/schema/ to match source structure
- Update all documentation references to reflect schema-based approach
- Align directory structure with "Parse, don't validate" pattern
- Schema-focused naming better reflects valibot usage and data
validation responsibility
