+++ src/schema/quiz.ts
